pyproject.toml
[project]
name = "toolbox-api"
version = "0.1.0"
description = "工具站后台服务，提供多种实用功能的 API 接口"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your@email.com" }
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "tortoise-orm>=0.22.0",
    "aiomysql>=0.2.0",
    "aiosqlite>=0.20.0",
    "redis>=5.2.0",
    "taskiq>=0.11.0",
    "taskiq-redis>=0.11.0",
    "aerich>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.aerich]
tortoise_orm = "app.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."